pyproject.toml
[project]
name = "dtest-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arize-phoenix>=13.21.0",
    "httpx>=0.28.0",
    "jupyterlab>=4.5.6",
    "langchain-core>=1.2.23",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.4",
    "langgraph-checkpoint-postgres>=3.0.5",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "openai>=2.30.0",
    "opentelemetry-exporter-otlp-proto-http>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "pandas>=2.3.3",
    "psycopg[binary]>=3.3.3",
    "psycopg-pool>=3.2.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.48",
    "grandalf>=0.8",
    "httpx-sse>=0.4.3",
    "nbformat>=5.10.4",
    "nbconvert>=7.17.0",
    "xgboost>=3.2.0",
    "lightgbm>=4.6.0",
    "shap>=0.49.1",
]
